Add a separate update cart from update order
The API for updating an order and cart are actually quite different in
the parts that they handle. Using a special update cart for just the
cart page makes the cart page actually work (recalculate totals, line
items with 0 quantity are deleted instead of just kept at 0
quantity). The API also deletes all shipments on every update which
makes it not suitable for usage within the checkout flow