Defer reading of text input values until write
This causes selenium (or very fast typists) to drop text input
characters while typing. To fix this, the events enqueue a function that
can retrieve the value (aka - a lazy evaluated value) to allow onChange
handler to have the latest text value when it needs to use it.

It's not perfect, but it does work....