Enqueue a navigate after signing in always
It will enqueue to the same page they are on and then cause the data to
be fetched. Additionally, api requests shouldn't go out if the user is
not actually signed in. We will guard here in effects for now but this
should be refactored into probably making a user a maybe monad so that
the requests are forced to not happen without a user