First pass at managing api requests in state
Separated out request keys from keypaths since maybe we can use them for
the cache too? Needs some more thinking there but just seemed nice to
have it separate for now.

The add-to-bag ugliness with placeholder request should be better when
the new api comes in but seems like it could be still valuable otherwise
-- represents a group of api requests that we can spin off of