Override yotpo styles on some new elements
Now that we have a paid account there are some new elements that need to
be styled.

We can also style inside of yotpo now with this paid plan
so maybe we should style there instead now? But then we have to do
additional effort to propogate styles from acceptance to production
which doesn't seem fun