Refactor api-requests from map to vector
For example, if navigating very fast, there can be two get-order
requests and on abort we need to abort both of them. Adds in request ids
to allow us to remove the correct one. Checking for a request can still
be done through request-key but uses the query method instead of a
direct get-in