Bundle builder only uses "premier" hair
It is tagged premier but the actual hair is deluxe. This is very
confusing. Unfortunately people still have hair that we don't sell
anymore in their carts but then those products show up as part of the
bundle builder. We need to eventually remove some skus from the bundle
builder nav taxonomy which would fix this but right now we are just
white/blacklisting in certain places for ease of undo-ing but this will
need to be cleaned up