Process paypal payments
Clicking on the paypal button will start the paypal checkout flow. Once
paypal redirects back to this pay, tell the backend to ask paypal
information about the order to process.

If that backend request succeeds, redirect the user to the order
complete page.

To reduce the likelihood of user drop-off in the redirect phase (eg -
user browser timeout), perform the redirect and order processing on the
server side of storefront.

Since this supports guest checkout via paypal, order complete no longer
requires a signed in user to view.