Add query param of loggedin to /checkout/address
To help analytics with the difference between the /checkout/address that
we redirect past vs the /checkout/address in which the user is actually
logged in. Waiting for analytics to decide what the more permanent
solution should be.

Navigate in effects then has to use the event/args
from the app-state and not the ones that were passed in, not sure how we
feel about this yet, something to think about some more