Reduce frequency of get-current-order calls
Now, storefront only calls get-current-order on:
1. App start
2. User Login
3. Cart navigate

This means that when a user is editing an order in two browsers, one
browser won't see changes from the other immediately.

This is mediated by a change to storeback, which ensures that add-to-bag
updates one single, shared current order, regardless of which browser
the request comes from.

So, if you add to cart in one browser, then add to cart in the other,
the second will see 2 items in the cart, instead of clobbering the first
browser's cart.