Simplify font files to always use woff
We don't support old IE which removes the need for .eot, and we don't
support old Safari which removes .svg and we don't support old Android
which removes .ttf (woff willl cover it) and the Extralight otf that was
present does not support modern IE, so needs to be a woff.

In the future as we get more woff2 files, we should probably update to
that as well with woff as a fallback.