Return 404 status codes when route not found
Instead of always returning index on any route and purely relying on
front end routing. We've held off on doing it because requires lifting
the routes as a cljc file so that the frontend and backend can share the
routing files.

Upgraded ring-defaults because the current version has version of
ring-headers that applies headers to empty responses, which coerces them
to actual ring respones -- preventing the not found handler from kicking
in. Similarly, removed the global not found route because routes.cljs
already handled the not found case as did handler.cj.

Upgraded org.clojure/tools.namespace to handle cljc file reloading.

Waiting on a better 404 page from Ryan than this static text of just
"Not found".

May be nice for handler.clj to just use bidi and remove compojure
entirely so there's not two routing libraries but at the moment the way
we are using bidi in handler.clj is not really for routing the different
endpoints but more of a presence test

Right now, /categories/hair/random will not return a 404, so the next
step is to handle these dynamic cases that will require a data fetch