Avoid using meta for passing app-version
Parsing now places the response body in a key ":body" on response.
Then, in merge-req-opts, :body is the only thing passed into the
handlers specified by each api function.
This enables us to put the app-version on a top level key, instead of
in meta.  In the future, if we want to use more response headers, we
could just glob them all into a headers key.

This commit also reintroduces null handling on getting the app-version