Restore defaults to just site routes
Can't rely on it too soon since it forces https and that causes havok on
cases like www.bob.mayvenn.com since the ssl cert can't cover two
levels. Since the defaults aren't brought in, we can't really rely on
the scheme yet. We probably need to split up the ring-defaults for this
app and put in the pieces in the right places since this is kind of unfortunate