Load main js file async not in acceptance/prod
It doesn't work in development since figwheel makes extensive use of
document.write whereas the compiled/optimized JS doesn't do that. Also,
for development, we'll have to use document ready since the script tag
is being brought in before the DOM is ready whereas in
acceptance/production, the document should already be ready since the
script is loaded async. There is some implications of menus and other
things not working until the script gets there that will have to be
solved later