Separate out seo tags to be cljc
Moved from opengraph tags only on the front end. It will be nice to have
these tags rendered server side since that is the last benefit of
prerender.io. Right now we always pick the default tags regardless of
the page we are on. Later, we can start picking different tags for
different pages