Inline stylesheet in production
Quite a few tradeoffs with this one

Pros:
- Less roundtrip requests
- No render blocking content (neither js or css)
- Prevents flash of unstyled content if just deferring the css
- Very large pagespeed improvement

Cons:
- No cache headers set so on refresh it will re-download styles.
Mitigated by single page app nature of site.
- Extra cycles for server to keep gzipping larger html that now includes
a lot of redundant content.

Future improvements: Could have nginx with it's mod page speed inline
for us which could have some improvements in keeping a gzipped version
of the css in memory. Can also parse the CSS and figure out the critical
parts to inline so that the whole stylesheet isn't there but the way
BassCSS works sort of means most of the stylesheet is going to be
relevant (one would imagine, haven't checked this).