Standardize built-component, and remove nesting
top-level was doing a component/build around all the main components.
This was necessary as we were transitioning to the component / query /
built-component paradigm.

This commit removes the outer component/build, which allows removal of
several nested levels of
  (component/create (component/build component ...))

In particular, built-component, wherever it is defined, now
returns component/build, not component/create.

Still TODO: use the component ns, even in cljs files that could use om
and sablono directly.