Allow mobile pixlee widget to properly sized
Potential Problem: When going from desktop to mobile layout and vice
versa, the widget disappears sometimes potentially because it was
originally hidden and pixlee set the iframe height to 0, and
resizeWidget doesn't fix that. We're unsure if this is worthy of
development time.

TODO: see if going from portrait to landscape mode and vice versa on
common devices results in pixlee widget to disappear.