Be more defensive about detaching wistia
If you click to open and close the video modal too quickly, wistia may
not be able to find the player, and will explode.

This might fix the bug where you can't close the player, especially if
the close button is directly above the open-video section and you click
both really quickly.