Don't fetch order in checkout flow
Unless that page was directly loaded. This works around a weird UX
issue where when getting to the confirm page, you may change the
shipping method but meanwhile the older version of the order loads and
switches the shipping method back.

We thought about preferring the user selected shipping method over the
order shipping method but that led to the inverted problem where the
shipping method is actually updated but the order totals are still
reflecting the previous shipping method choice.