Remove dead code from flash handling
We used to have more specific handling of when to dismiss the
flash but the refactor of error messages removed it. We've been showing
the flash on next (usually return nav) page and dismissing for about
two months. The handling is simplified because we don't store the
navigation message at all. We clear flashes on all navigate events. This
works because we navigate before we flash in all effects. For an
instance of that check out how we navigate away from the login page when
a user is already signed in.