Try to fix exception when extracting query from url
We get a lot of errors like, "Cannot read property 'string' of
undefined", for example
https://app.bugsnag.com/mayvenn/storefront-cljs/errors/57c0a7c054bf78e632e717c4?filters%5Bevent.since%5D%5B%5D=30d&filters%5Berror.status%5D%5B%5D=open&filters%5Bevent.message%5D%5B%5D=Cannot%20read%20property%20%27string%27%20of%20undefined
Usually this happens in IEquiv. (Sometimes it's comparing booleans, or
other primitives, but let's consider strings in this case.)

Could it be that we are trying to compare a string with the JS value
'undefined'? Probably undefined doesn't have a conversion to string,
which could make IEquiv blow up.

If that's the case, we need to find out how 'undefined' makes it into
our code. If you just evaluate js/undefined, it returns nil. And we can
definitely compare nil and a string. But perhaps, in some scenarios,
undefined isn't actually nil.

One obvious potential source of undefined is js interop code. This fix
addresses one interop case, which throws errors about once a day.
Perhaps js/location.href is sometimes undefined.

If this fix removes occurrences of the error (we'll have to look for
exceptions coming from the routes ns), we should be more careful
about casting js interop values to clojurescript values.

Further references:
https://groups.google.com/forum/#!msg/clojure/f0q9kHwMQ7Q/EoJ1QU299BMJ
http://stackoverflow.com/questions/22046889/clojurescript-to-check-if-a-property-exists-in-a-js-object