Fetch SVG sprite from an external resource
Instead of inlining it into the page. Decreases initial document size
(and JS) and leverages browser resource caching.

Won't work with IE, does work with Edge. Some reports of it not working
in older Android. If this is a problem, we could inline the SVG with
AJAX: https://css-tricks.com/ajaxing-svg-sprite/.