Move fill-dark-silver to inc/dec counter svgs
Chrome doesn't seem to always render with the fill value provided when
the fill-dark-silver is set on a parent element to the svg. Multiple
reloads of the page produces different behaviors (sometimes black, the
svg's value, and sometimes it's dark-silver, the intended value). It
seems to have been fixed by moving the classes directly on the svg.

There's no color customizations other than dark-silver for its usages.