Parse nav messages identically on server & client
It is important that both client and server generate the same nav
messages, because otherwise pageviews can be double counted.

The app-start event logs a pageview. Usually the nav event fires, but
does not log a pageview because the nav message has not changed from
what the server set it as.

However, if the client and server generate different nav messages, the
nav event can also trigger a pageview.

This fix ensures that client and server use the same code to generate
nav messages.