Stop double-counting pageviews, and fix flash
There are two fixes here.

First, #354ea92d broke the flash on sign-in/sign-up and other pages that
redirect and show a flash. Those pages were depending on the redirection
navigate event happening (and clearing the flash) before the flash was
set (synchronously) by a `(handle-message events/flash-show-success)`.
Now that the redirection navigate event happens asynchronously, it was
clearing the flash immediately after it was set. Now, when redirecting,
trigger `(handle-message events/flash-later-show-success)` to show the
flash on the redirect page. You can still `(handle-message
events/flash-show-success)` to show the flash on the current page.

The pageviews fix stops pages that redirect from counting as a pageview.
It does this by letting the redirect effect trigger a synchronous
transition to store the fact that the first page is redirecting. Though
`core/handle-message` allows effects to trigger synchronous transitions,
we should not abuse this power.

Finally, the nav message is removed from the server side app state
before passing it on to the client. This lets us remove the pageview
tracking on app-start, thus not always counting an initial page as a
pageview, even if that page is destined to redirect.