Allow repeated calls to convert assignVariation
This was introduced when I was thinking convert would be our cache for
whether a user had been bucketed into the address-login experiment.
Since we can't count on convert to be loaded early enough, I switched to
the current algorithm in experiments/bucket-for, which doesn't need
convert.

Since it's not clear whether convert actually cares if we tell them
multiple times, this removes the cache that prevents repeated calls. It
had become a premature optimization essentially.