Stop trying to inform convert about bucketing
The `assignVariation` helper Convert exposes only works if you call it
before their script tags load. For this experiment, that doesn't help
us. We need to trigger the variation bucketing only when the user clicks
the Checkout button. At that point we know whether they're logged in (in
which case they don't get bucketed into either experiment) or not (in
which case they should get bucketed into original or variation).