Restructure pixlee data, to store images in one big database
Images used to be stored at a few different keypaths, based on which
album they were in. This puts them all together, and keeps track of
which images are in which albums separately.

This will make it possible to fetch an individual image when the look
details page is loaded directly. This is not just a performance win. It
will also let us stop assuming that all detailed looks can be fetched by
fetching the mosaic album. This is necessary because we are about to
have look details that are from category pages, not the mosaic.

NOTE: we are now using the album_photo_id in the look detail url. This
id will be used to fetch an individual image from the pixlee
/api/v2/media API. Unfortunately, this means all existing look details
URLs will be broken.