Attempt to separate event handling multimethods
We have placed the multimethods for events in the component file.
We think this will possibly have multiple future benefits. A) This will
pave the way so that we can handle events on the server, collapsing
the divergence in backend and frontend data handling. We also think
that this could result in dynamically loaded frontend modules reducing
the size of the initial js payload.

[#nodeps]