Switch redirects to shop.mayvenn.com to a 301
For SEO ownership transfer, mayvenn.com and www.mayvenn.com should redirect to
shop.mayvenn.com through a 301 instead of a 302. There seems to be no harm in
redirecting internal.mayvenn.com to shop.mayvenn.com with a 301 as we don't want
that store to exist.

Change the https redirect test to have a stylist store (bob) instead of no store
since the redirect order is from no stylist store -> stylist store and so
technically it wasn't solely testing the HTTPS redirect.