Update feature blocks for Straight and Body Wave bundle deals [#nodeps]
NOTE: these assets are larger than what we currently have in these
feature blocks.  I played with them in my inspector and it just worked
without interrupting the rest of the page layout.  we should double
check to make sure that is correct.

basically we want square blocks on mobile and larger landscape blocks
on desktop.