Fix double encode/decode cookies on server side
wrap-defaults implies wrap-cookies and the middleware isn't aware if it has
already encoded/decoded values in the request or response.

In order to not break all existing cookies, storefront now only encodes once for
cookies and continues to double-decode cookie values when reading since most of
our cookie values do not require special values.

Of the cookies that do need some special handling of single-encode,
double-decode, a simple 's/ /+/g' is good enough. email and order token are the
only two values that need such migration, but email was already broken so it is
being left-as-is.

This double serialization behavior only applied to storefront server and not the
client side.

This is not as dangerous a bug as it seems from the outset because the server
typically passes parsed EDN data to the frontend and the client then uses
single-encoding to rewrite to the cookies key again. There is only a handful of
cases where cookies are used to share data from server to client:

- server-only endpoints that redirect to a frontend page
- the email address cookie page views

Co-authored-by: Justin Watt <justinwatt@gmail.com>