Show warning when passing nil values into money formatters
We shouldn't try to format nil ever.  This leads to things like $0.nan and
whatnot.  We also don't want to just assume 0 without telling anyone, and we
also did not want to break rendering of this.  So we compromised and assume 0
but throw an exception.  In cljs we immediately catch it to prevent breaking the renderer.

Co-authored-by: Ellie Winn <ellie@mayvenn.com>