Catch exceptions when fetching from contentful
Repro:

If contentful had connection issues, storefront would stop polling. A simple way
to verify this behavior:

  (let [pool (at-at/mk-pool)]
    (try
      (at-at/every 1000
                  #(do
                      (println "HI")
                      (throw (Exception. "NO")))
                  pool)
      (println "Sleeping")
      (Thread/sleep 3000)
      (println "Done Sleeping")
      (finally
        (at-at/stop-and-reset-pool! pool))))

   ;;; prints the following:
   ;; Sleeping
   ;; HI
   ;; Done Sleeping

Reason:

This is due to java still raising SocketExceptions when a timeout occurs. This
exception isn't caught by us and propagates to at-at, which stops the polling
when it sees an exception.

Fix:

The fix is to explicitly catch IOExceptions (parent of SocketException) as a
failure when making a request to contentful and allow the retry mechanisms to
continue to work. There's also a try-catch around the entire call to ensure it
has a fallback-capture and report to bugsnag incase any other exceptions are
raised.

[Fixes #160270600]