[Experiment]: Insert app script tags in body
Do not deploy to prod without talking to Jeff first.

This is an experiment to see the performance impact of this change. A local
production compile *evaluates* in half the time locally vs acceptance or
production. This seems to indicate that some different in how the script is
loaded makes this fast for some reason.

This attempts to insert the JS files the same way as development to see if that
generates a difference or not.