SEO: Lazy load carousel images
Use defer-ucare-img on ICP pages. Restrict ICP/Category product image sizes

screen-aware-component now has a way to specific trigger a loaded state. This
is useful if the inner component loads images and wants to trigger a rerender
when that image loads. Currently it's used in defer-ucare-img, although it might
be better to extract to its own component some day.

To help make the carousel image loading not as awkward, a spinner and a quick
fadein is used. Open to potential design changes since it was was quick to
implement.