Drop post-purchase matching of stylists
Warning, includes significant refactors to all of matching.

A model was created to represent the flow of Stylist Matching.

Read model: stylist-matching.core/stylist-matching<-

This model is built from event sourcing events representing:

- parameter constrained, "location, service, etc"
- prepared, "Search should be performed"
- searched, "Search is performed"
- resulted, "Search was performed and there are results"

The end result of this is fewer confusingly named keypaths, and
concretization of the behavior. We hope this makes it easier to
grok and modify.

Further work is needed to model selecting a stylist for the order
as this concept is the bridge between matching and ordering.

Co-authored-by: Corey Reece <corey@mayvenn.com>
Co-authored-by: Jonathan Harford <jonathanh@mayvenn.com>