Define adventures
Adventures are a kind of visual layer.

They represent multi-screen experiences to arrive at a goal instead
of have a lot of UI on a single screen. The trade-off upside is to reduce
cognitive load at the same time allowing a lot of flexibility from the
user. The down-side is that they don't allow as many degrees of freedom
at once, so explorability at a glance is lower.

They are traditionally called a wizard or user flow.

[#178308328]

Co-authored-by: Corey Reece <corey@mayvenn.com>
Co-authored-by: Justin Watt <justinw@mayvenn.com>