Top Stylist: HACK - fix for direct loading stylist results page
The stylist searching is tied directly to the dom form input element
so it was not searching for stylist on load because the input had not
mounted. Rather than not mounting the stylist search location input,
we opted for simply mounting and hiding it. Probably this search
should be driven from flow events more directly.

[#178474227]

Co-authored-by: ~Stella Clemens <stella@mayvenn.com>
Co-authored-by: Jonathan Harford <jonathanh@mayvenn.com>