HACK: Enable features handles both string and map entries
Evidently the transitions for enable-feature takes both LD FFs and convert
experiments. This hasn't been a problem in the past when both were acting as
boolean flag stores, but now that LD has variants (and its data store is a map
rather than a vector) this has caused some problems. Likely this will cause some
issues with HEAT...

This is marked as a hack because we simply shouldn't be storing LD and convert
FFs in the same app-state location, it's not safe to treat them as transparently
interchangeable.