SMS: Update opt-in according to telecom feedback
- Opt-in must not be a checkbox
- Include HELP/STOP notice

The reasoning given is that opt-in should be resistant to people who
don't read labels and assume it is an agreement to check / required for
purchase.

Design-System: Reduce right-margin from radio button to label

Visually couples the radio button to the label better when there are two
radio buttons horizontally laid out, as per SMS opt-in.

Proposed auditor design wanted more of a table layout:

------------------------------------------------
 Do you want to receive order updates via text?

  Yes  No
  ( ) (x)  ... send me order updates.
  ( ) (x)  ... send me marketing messages.
------------------------------------------------

However, this makes accessibility suffer (no label attached to radio
button). But now spacing between label and other radio buttons is now
important.

Implemented design (Desktop):

-----------------------------------------------------
 Do you want to receive order updates via text?

  ( ) Yes  (x) No   ... send me order updates.
  ( ) Yes  (x) No   ... send me marketing messages.
-----------------------------------------------------

Implemented design (Mobile):

-----------------------------------------------------
 Do you want to receive order updates via text?

                  ( ) Yes  (x) No
            ... send me order updates.

                  ( ) Yes  (x) No
           ... send me marketing messages.
-----------------------------------------------------

Centering provides easier tap targets (edge of screen has lower number
of capacitors to detect finger presses).

The ellipsis in the copy is to make this form of the design easier to
understand without a more complicated reordering of the layout.

This above implemented design was approved by the auditors.