PDP Redesign: add Components Library page
We grabbed 5 example components to add based solely on ease of faking their
queries. Also, the video one is a popup, so it kindof disables the page.

Next steps are to make changes to the accordion component so that it matches the
specifications and then merge the accordion with the tabbed information
component.

Co-authored-by: Jonathan Harford <jonathanh@mayvenn.com>
Co-authored-by: Stella Clemens <Stella@mayvenn.com>